fix: update manifest and link controller behaviour
Add asynchronous connect to LTE. This means that the device
will initially configure the modem upon boot and
go straight to GPS search. When the modem connects to LTE it
the device will perform its usuall cloud synchronizatio routine
etc.. This behaviour allows the device to obtain and buffer GPS
fixes before a connection to LTE if no LTE network coverage
is present. Closes #45

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>