feat: store all data in circular buffers
All data published to cloud is now stored in
circular buffers. In the end of a cloud publication cycle
the cicular buffers are checked if they contain
unpublished(queued) data. If this is the case the
queued entries are bundled togheter in corresponding
objects and published to a custom batch topic.
Closes #32.

The size of these cicular buffers and how many data
entries that are bundled togheter in each batched message
are statically configurable options.

Signed-off-by: Simen S. Røstad <simen.rostad@nordicsemi.no>